pyproject.toml
[project]
name = "microlane"
version = "1.0.0"
description = "Evaluating how well modern lane detection models hold up when deployed on small-scale 1/10 RC cars"
authors = [
    {name = "suyog prasai",email = "suyog224356@gmail.com"}
]
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "docker (>=7.1.0,<8.0.0)",
    "numpy (>=2.4.4,<3.0.0)",
    "opencv-python (>=4.13.0.92,<5.0.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "click (>=8.3.3,<9.0.0)",
    "filetype (>=1.2.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "microlane" },
    { include = "scripts" }
]

[tool.poetry.scripts]
microlane = "scripts.microlane:main"